+++ hile.h
@@ -65,8 +65,8 @@ typedef struct {
   uint32_t pos;
   char reference_base;
   hile_basestrand_t *bases;
-  uint16_t n;
-  uint16_t cap;
+  uint32_t n;
+  uint32_t cap;
   uint8_t *bqs;
   uint8_t *mqs;
   char **read_names;
